requirements.txt
anthropic==0.3.0
ebooklib==0.17.1
rpy2==3.5.11
pytest==7.3.1
python-dotenv==1.0.0
